// src/providers/cremedelamer/types.ts
export interface Sku {
  path: string
  MISC_FLAG: number
  SKU_REPLACEMENT_2: string
  LARGE_ALT_IMAGES: string[]
  SKIN_TONE: string[]
  SKU_IMAGE_1?: any
  VTO_FLAG_ONLINE?: any
  SKU_BASE_ID: number
  formattedUnitPrice: string
  MEDIUM_IMAGE: string[]
  rs_sku_pc: string
  RELEASE_DATE?: any
  HEX_VALUE: string
  IMAGE_LARGE: string[]
  SKIN_TONE_CHECKBOXES?: any
  ENGRAVABLE?: any
  REGION_ID: number
  DISPLAY_ORDER: number
  IMAGE_MEDIUM: string[]
  IMAGE_RESPONSIVE: string
  SMOOSH_PATH_STRING: string
  WWW_PCODE: any[]
  FINISH?: any
  isLimitedRemaining: number
  REFILLABLE: number
  ATTRIBUTE_COLOR_FAMILY?: any
  IS_ENGRAVABLE?: any
  UOM?: any
  PRODUCT_ID: string
  SHADENAME: string
  PRICE: number
  SMOOSH_DESIGN?: any
  PRODUCT_CODE: string
  LIFE_OF_PRODUCT: number
  SKIN_TONE_TEXT?: any
  NUMBER_OF_ALT_IMAGES_MEDIUM: number
  SKU_REPLACEMENT_1: string
  SHADE_DESCRIPTION: string
  rs_sku_availability: number
  SKIN_TYPE_TEXT?: any
  BRAND_ID: number
  isOrderable: number
  SKIN_TONE_REFINEMENT?: any
  SKIN_TYPE: number
  PRODUCT_SIZE: string
  formattedPrice: string
  rs_sku_image: string[]
  SKU_ID: string
  SIZE_DESCRIPTION?: any
  UNDERTONE_FINDER: string
  formattedInstallmentPrice?: any
  rs_sku_limited_remaining: number
  PARENT_CAT_ID: string
  NUMBER_OF_ALT_IMAGES_LARGE?: any
  SKIN_TONE_1: string
  HEX_VALUE_STRING: string
  MISC_FLAG_TEXT: string
  isShoppable: number
  DEFAULT_SKU: number
  COLORGROUPING?: any
  LARGE_IMAGE: string[]
  rs_sku_price: number
  UNDERTONE?: any
  SUPPRESS_SEARCH?: any
  WWW_PCODE_2?: any
  IMAGE_SMALL: string
  NUMBER_OF_ALT_IMAGES: number
  UNDERTONE_CHECKBOXES?: any
  UPC_CODE: string
  STRENGTH?: any
  isPreOrder: number
  INVENTORY_STATUS: number
  WWW_PCODE_1?: any
  UNIT_SIZE?: any
  DISPLAY_STATUS: number
  formattedTaxedPrice: string
  shadeMenuLabel: string
  isComingSoon: number
  video?: {
    videoImage?: string
    'provider-youtube'?: boolean
    youtube_id?: boolean
  }
}

export interface Category {
  REGION_ID: number
  DESCRIPTION: string
  MNEMONIC: string
  CAT_BASE_ID: number
  BRAND_ID: number
  RGN_SUBLINE?: any
  CATEGORY_ID: string
  META_DESCRIPTION?: any
  PARENT_CAT_ID?: any
  CATEGORY_NAME: string
  DISPLAY_STATUS: number
}

export default interface LaMerProduct {
  path: string
  DESCRIPTION?: any
  RATING_RANGE?: any
  BENEFIT_STATEMENT?: any
  META_DESCRIPTION: string
  ILN_FORMULATED_WITHOUT?: any
  MEDIUM_IMAGE: string[]
  PROD_SKIN_TYPE_TEXT?: any
  PRODUCT_TABBED_DATA?: any
  MNEMONIC?: any
  SKIN_CONCERN_ATTR?: any
  worksWith: any[]
  shaded: number
  IMAGE_LARGE: string[]
  defaultPath: string
  PROD_CAT_DISPLAY_ORDER: number
  MAINCAT_CHECKBOXES?: any
  FORMULA?: any
  SKIN_CONCERN_LABEL?: any
  REGION_ID: number
  SMALL_IMAGE: string
  IMAGE_RESPONSIVE: string
  RATING_IMAGE: string
  sized: number
  SD_FINISH_CHECKBOXES: string[]
  PROD_CAT_IMAGE_NAME: string
  SKIN_TYPE_CHECKBOXES?: any
  PRODUCT_ID: string
  THUMBNAIL_IMAGE: string
  BENEFIT_CHECKBOXES?: any
  RECOMMENDED_COUNT?: any
  FINISH_CHECKBOXES: string
  skus: Sku[]
  SKIN_CONCERN_1?: any
  PROD_RGN_NAME: string
  rs_default_pc: string
  rs_default_rating: number
  RECOMMENDED_PERCENT: number
  ATTRIBUTE_DESC_1: string
  TOTAL_REVIEW_COUNT: number
  SKIN_CONCERN_2?: any
  BRAND_ID: number
  AVERAGE_RATING: number
  rs_default_image: string[]
  ONLY_RATINGS_COUNT?: any
  skinTone: number
  ATTRIBUTE_BENEFIT?: any
  rs_default_count: number
  PRODUCT_USAGE?: any
  RGN_SPECIAL_NAME?: any
  url: string
  rs_default_desc: string
  rs_default_name: string
  skinType: number
  PARENT_CAT_ID: string
  SD_FORM_CHECKBOXES: string[]
  DESCRIPTION_IPAD?: any
  NUMBER_OF_ALT_IMAGES_LARGE?: any
  ATTRIBUTE_COVERAGE?: any
  rs_default_url: string
  DEFAULT_CAT_ID: string
  isShoppable: number
  rs_default_currency: string
  FORM_CHECKBOXES: string
  LARGE_IMAGE: string[]
  SKINTYPE_DESC?: any
  SUB_LINE: string
  PROD_SKIN_TYPE?: any
  PROD_IMAGE_1?: any
  COVERAGE_OPTIONS_CHECKBOXES: string
  SUPPRESS_SEARCH?: any
  SHORT_DESC: string
  PROD_RGN_SUBHEADING?: any
  DISPLAY_STATUS: number
  category: Category
  PROD_BASE_ID: number
  lowest_sku_price: string
  lowest_sku_price_taxed: string
  lowest_sku_price_unformatted: number
  defaultSku: Sku
  shaded_sku_count: number
  filtered_skintone: string[]
  filtered_skintone_filter: boolean
  has_shade_filter: boolean
  filtered_undertone: string[]
  filtered_undertone_filter: boolean
  MISC_FLAG_TEXT: string
  displayOrder: number
  displayStatus: number
  MISC_FLAG: number
  SKU_REPLACEMENT_2: string
  LARGE_ALT_IMAGES: string[]
  SKIN_TONE: string[]
  SKU_IMAGE_1?: any
  VTO_FLAG_ONLINE?: any
  SKU_BASE_ID: number
  formattedUnitPrice: string
  rs_sku_pc: string
  RELEASE_DATE?: any
  HEX_VALUE: string
  SKIN_TONE_CHECKBOXES?: any
  ENGRAVABLE?: any
  DISPLAY_ORDER: number
  IMAGE_MEDIUM: string[]
  SMOOSH_PATH_STRING: string
  WWW_PCODE: any[]
  FINISH?: any
  isLimitedRemaining: number
  REFILLABLE: number
  ATTRIBUTE_COLOR_FAMILY?: any
  IS_ENGRAVABLE?: any
  UOM?: any
  SHADENAME: string
  PRICE: number
  SMOOSH_DESIGN?: any
  PRODUCT_CODE: string
  LIFE_OF_PRODUCT: number
  SKIN_TONE_TEXT?: any
  NUMBER_OF_ALT_IMAGES_MEDIUM: number
  SKU_REPLACEMENT_1?: any
  SHADE_DESCRIPTION?: any
  rs_sku_availability: number
  SKIN_TYPE_TEXT?: any
  isOrderable: number
  SKIN_TONE_REFINEMENT?: any
  SKIN_TYPE: number
  PRODUCT_SIZE: string
  formattedPrice: string
  rs_sku_image: string[]
  SKU_ID: string
  SIZE_DESCRIPTION?: any
  UNDERTONE_FINDER: string
  formattedInstallmentPrice?: any
  rs_sku_limited_remaining: number
  SKIN_TONE_1: string
  HEX_VALUE_STRING: string
  DEFAULT_SKU: number
  COLORGROUPING?: any
  rs_sku_price: number
  UNDERTONE?: any
  WWW_PCODE_2?: any
  IMAGE_SMALL: string
  NUMBER_OF_ALT_IMAGES: number
  UNDERTONE_CHECKBOXES?: any
  UPC_CODE: string
  STRENGTH?: any
  isPreOrder: number
  INVENTORY_STATUS: number
  WWW_PCODE_1?: any
  UNIT_SIZE?: any
  formattedTaxedPrice: string
  shadeMenuLabel: string
  isComingSoon: number
}